requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
pydantic-settings==2.6.0
pymongo==4.9.1
websocket-client==1.8.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1
